F/auth roles

Add authentication and authorisation

everything is now totally locked down!

See merge request !5

+++ server/methods.coffee
@@ -20,7 +20,9 @@ sshexec = (id, command) ->
 
 Meteor.methods
   restartGalicaster: (id) ->
-    sshexec id, 'killall python2'
-
+    if isUserAuthorised Meteor.userId(), ['admin', 'control-rooms']
+      sshexec id, 'killall python2'
+      
   rebootMachine: (id) ->
-    sshexec id, 'sudo reboot'
+    if isUserAuthorised Meteor.userId(), ['admin', 'control-rooms']
+      sshexec id, 'sudo reboot'
